For the sake of consistency updated the camel case formatting in the parameter names defined in the hardhat.config file
